refactor: 서비스들이 관리자 권한에 따라 분리되지 않도록 수정
