+++ src/main/java/roomescape/presentation/controller/admin/ReservationTicketAdminController.java
@@ -8,19 +8,21 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.service.ReservationTicketAdminService;
-import roomescape.dto.request.ReservationAdminRegisterDto;
+import roomescape.application.service.ReservationTicketService;
+import roomescape.dto.LoginMember;
+import roomescape.dto.request.ReservationTicketRegisterDto;
 
 @RestController
 @RequestMapping("/admin/reservations")
 @RequiredArgsConstructor
 public class ReservationTicketAdminController {
 
-    private final ReservationTicketAdminService reservationTicketAdminService;
+    private final ReservationTicketService reservationTicketService;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void addReservation(@RequestBody @Valid ReservationAdminRegisterDto reservationAdminRegisterDto) {
-        reservationTicketAdminService.saveReservation(reservationAdminRegisterDto);
+    public void addReservation(@RequestBody @Valid ReservationTicketRegisterDto reservationAdminRegisterDto,
+                               LoginMember loginMember) {
+        reservationTicketService.saveReservation(reservationAdminRegisterDto, loginMember);
     }
 }
